refactor: put unwindDoomed beside the truncation it describes
